refacot(rpc): refactor `rpc/eth` package
